requirements.txt
pyTelegramBotAPI==4.5.1
pyowm==3.3.0
googletrans==3.1.0a0
beautifulsoup4==4.11.1
pdfplumber==0.6.2
gTTS==2.2.4
art==5.6
python-dotenv==1.0.0

